DES_IP/Protocols/requirements.txt
# Requirements for UART communication with Altera DE2
pyserial>=3.5
